refs #30885 - routing change for ghostworkstations
The former route was intended for the frontend, not the API. But we need
a seperate route for rights managment to allow setting this parameter
without granting access to change scopes.
Additionally introduced a check if the useraccount has access to the
accessed scope and throw an exception on a try to change the
ghostworkstations of alien scope.